+++ CHANGES.rst
@@ -14,7 +14,11 @@ Changes:
 
 Fixes:
 ------
-- No change.
+- Fix reported ``links`` by processes nested under a provider ``Service``.
+  Generated URL references were omitting the ``/providers/{id}`` portion.
+- Fix documentation referring to incorrect setting name in some cases for WPS outputs.
+- Fix strict XML parsing failing resolution of some remote WPS providers with invalid characters such as ``<``, ``<=``
+  within process description fields. Although invalid, those easily recoverable errors will be handled by the parser.
 
 `4.0.0 <https://github.com/crim-ca/weaver/tree/4.0.0>`_ (2021-09-21)
 ========================================================================